[fix]ユーザーアイコンとローディングのコンポーネント化
